Update: Video タグを分岐するコンポーネントを作成し、より汎用的に動画を再生できるように修正
