+++ src/components/atoms/AVideoSource/index.ts
@@ -1,2 +1,2 @@
-export { AVideo } from "./component";
+export { AVideoSource } from "./component";
 export type { Props as AVideoProps, SrcType as AVideoType } from "./component";
