+++ src/components/atoms/AVideoStream/hooks.tsx
@@ -1,4 +1,4 @@
-import React, { createRef } from "react";
+import { createRef } from "react";
 
 export const useAVideoStream = () => {
   const videoRef = createRef<HTMLVideoElement>();
