+++ src/components/molecules/AVideo/index.ts
@@ -1,0 +1,2 @@
+export { AVideo } from "./component";
+export type { VideoSource } from "./component";
