+++ src/components/molecules/ThreeScreenVideo/hooks.ts
@@ -1,12 +1,3 @@
-import { AVideoType } from "../../atoms/AVideo";
-import { AVideoStreamType } from "../../atoms/AVideoStream";
-
 export const useThreeScreenVideo = () => {
-  const isMediaStream = (
-    src: AVideoType | AVideoStreamType
-  ): src is MediaStream => typeof src !== "string" && src !== null;
-
-  return {
-    isMediaStream,
-  };
+  return {};
 };
